Move imported code to the right location
This commit only moves the code, but it doesn't adapt it to this repo
layout yet.
